assistant: Extract `ContextEditor` to its own module
