+++ crates/assistant/src/assistant.rs
@@ -2,6 +2,7 @@
 
 pub mod assistant_panel;
 mod context;
+mod context_editor;
 pub mod context_store;
 mod inline_assistant;
 mod patch;
